Move auth- and api-instance to new Procosys-context
